remove unused code
This removes the old `token` and `email` field from the
responses. `email` should now be added by the editors
as a part of their `form_json`. This allows them to ask
for the address at any point in the questionaire.
Since we no longer allow users to edit their responses
we remove the `token` field entirely.