disable django-restframework's html view
Django-Rest-Framework comes with a browsable API which renders information
about the API when users are making request with their web
browser directly instead of via AJAX.
As we do not expect anyone to be using our API right
now and do not include the required templates for these
views anyways it makes sense to just turn it off for now.

We might to turn it on back later as it can be quite useful.